refactor: Move steps visualization to widgets.steps_widgets
