--- src/visualization/associations/__init__.py
@@ -1,2 +1,0 @@
-from .a_priori_vis import APrioriStepsVisualization
-from .a_priori_canvas import APrioriScatterPlot, APrioriGraphPlot, APrioriGauge
