--- src/visualization/classification/__init__.py
@@ -1,1 +1,0 @@
-from .extra_trees_vis import ExtraTreesStepsVisualization
