--- src/visualization/clustering/__init__.py
@@ -1,2 +1,0 @@
-from .k_means_vis import KMeansStepsVisualization, KMeansCanvas
-from .gmm_vis import GMMStepsVisualization, GMMCanvas
