+++ src/widgets/steps_widgets/algorithm_vis.py
